new test.rs file with fixes for stake_eth function

+++ contracts/chrysalis/src/contract_interface.rs
@@ -12,7 +12,7 @@ pub trait ChrysalisContractTrait {
     fn initialize_contract(env: Env, staked_token_address: Address, steth_address: Address);
 
     /// Returns the current staked amount for the user.
-    fn get_stake_amount(e: Env, user: Address) -> i64;
+    fn get_stake_amount(e: Env, user: Address) -> i128;
 
     /// Returns the current balance of stETH for the user.
     // fn get_steth_balance(e: Env, user: Address) -> i64;
@@ -26,7 +26,7 @@ pub trait ChrysalisContractTrait {
     /// * `e` - An instance of the `Env` struct.
     /// * `user` - The address of the user staking the tokens.
     /// * `amount` - The amount of the custom token to stake.
-    fn stake_eth(e: Env, user: Address, amount: i64);
+    fn stake_eth(e: Env, user: Address, amount: i128);
 
     /// Unstakes a specified amount using stETH, burns the stETH, and returns the original tokens.
     /// 
@@ -35,10 +35,10 @@ pub trait ChrysalisContractTrait {
     /// * `e` - An instance of the `Env` struct.
     /// * `user` - The address of the user unstaking the tokens.
     /// * `amount` - The amount of stETH to use for unstaking.
-    fn unstake_eth(e: Env, user: Address, amount: i64);
+    fn unstake_eth(e: Env, user: Address, amount: i128);
 
-    fn claim(env: Env, user: Address)-> i64;
+    fn claim(env: Env, user: Address)-> i128;
 
-    fn vclaim(env: Env, user: Address)-> i64;
+    fn vclaim(env: Env, user: Address)-> i128;
     
 }
